export `image::RgbaImage`
instead of create and export type `Buffer`

+++ filter/benches/filter.rs
@@ -15,7 +15,7 @@ fn adaptive_threshold(b: &mut Bencher) -> Result<(), Box<dyn Error>> {
     let path = PathBuf::from(IMG);
     let img = image::open(path)?.into_rgba8();
 
-    b.iter(|| filter::adaptive_threshold(&img, RADIUS, FACTOR));
+    b.iter(|| filter::adaptive_threshold(&img, RADIUS as usize, FACTOR));
     Ok(())
 }
